+++ dds/src/implementation/dds_impl/mod.rs
@@ -6,7 +6,6 @@ pub mod dds_data_reader;
 pub mod dds_data_writer;
 pub mod dds_domain_entity_list;
 pub mod dds_domain_participant;
-pub mod dds_domain_participant_factory;
 pub mod dds_entity_list;
 pub mod dds_publisher;
 pub mod dds_subscriber;
