+++ dds/src/implementation/dds_impl/node_domain_participant.rs
@@ -1,13 +1,10 @@
 use crate::{
     builtin_topics::{ParticipantBuiltinTopicData, TopicBuiltinTopicData},
     domain::{
-        domain_participant_factory::DomainId,
+        domain_participant_factory::{DomainId, THE_DDS_DOMAIN_PARTICIPANT_FACTORY},
         domain_participant_listener::DomainParticipantListener,
     },
-    implementation::{
-        dds_impl::dds_domain_participant_factory::THE_DDS_DOMAIN_PARTICIPANT_FACTORY,
-        rtps::types::Guid, utils::node::ChildNode,
-    },
+    implementation::{rtps::types::Guid, utils::node::ChildNode},
     infrastructure::{
         condition::StatusCondition,
         error::{DdsError, DdsResult},
