+++ dds/src/implementation/dds_impl/node_user_defined_data_reader.rs
@@ -1,7 +1,7 @@
 use crate::{
     builtin_topics::PublicationBuiltinTopicData,
+    domain::domain_participant_factory::THE_DDS_DOMAIN_PARTICIPANT_FACTORY,
     implementation::{
-        dds_impl::dds_domain_participant_factory::THE_DDS_DOMAIN_PARTICIPANT_FACTORY,
         rtps::{stateful_reader::RtpsStatefulReader, types::Guid},
         utils::{
             node::ChildNode,
